back-end/flows/requirements.txt
beautifulsoup4 == 4.11.1
pandas == 1.4.2
psycopg2-binary == 2.9.2
pydantic == 1.9.0
python-dotenv == 0.19.2
requests == 2.26.0
sqlalchemy == 1.4.27